Python_basic/Requests_lib/Task2/requirements.txt
request==2.21.0
python-dotenv==0.10.3